requirements.txt
APScheduler==3.6.3
cachetools==4.2.2
certifi==2024.2.2
CurrencyConverter==0.17.20
psycopg2-binary==2.9.9
python-dotenv==1.0.1
python-telegram-bot==13.7
pytz==2024.1
six==1.16.0
tornado==6.4
tzlocal==5.2
